test01

+++ src/main/java/TheatreTicketBookingSystem/factories/VIPFactory.java
@@ -10,7 +10,7 @@ import java.util.Map;
 public class VIPFactory {
     public static VIP getVIP(Map<String, Object> values){
         VIP factoryVIP = new VIP.Builder()
-                .seat_id((Integer)values.get("seat_id"))
+                .seat_id((Long)values.get("seat_id"))
                 .price((Integer)values.get("price"))
                 .class_name((String)values.get("class_name"))
                 .build();
